% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MLmodels_fct_biosigner.R
\name{assessPerformance_Biosigner}
\alias{assessPerformance_Biosigner}
\title{MLmodels biosigner function}
\usage{
assessPerformance_Biosigner(modelList, dataTest)
}
\arguments{
\item{modelList}{Models list computed and returned inside
runSVMRFmodels_Biosigner() results.}

\item{dataTest}{List of new omic datasets and samples for same omics that
training to test model performances. Returned by splitDatatoTrainTest().}
}
\value{
List of performances for svm and rf models (confusion matrices).
}
\description{
A biosigner models assessing function.
to display by confusion matrices for SVM and RF models.
}
\examples{
data("omic2", package = "multiSight")
splitData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splitData$data.train
data.test <- splitData$data.test

#biosignerRes <- runSVMRFmodels_Biosigner(data.train)
data("biosignerRes", package = "multiSight") 
biosignerModels <- biosignerRes$model #list of SVM/RF models for each omic.
biosignerFeats <- biosignerRes$biosignature #selected features for each omic.
perfBiosigner <- assessPerformance_Biosigner(biosignerModels, data.test)
perfBiosigner$svm$rnaRead # perf for SVM for rnaRead data block.
perfBiosigner$rf$rnaRead # perf for RF for rnaRead data block.

}
