% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MLmodels_fct_diablo.R
\name{assessPerformance_Diablo}
\alias{assessPerformance_Diablo}
\title{MLmodels diablo function}
\usage{
assessPerformance_Diablo(splsdaModel, dataTest)
}
\arguments{
\item{splsdaModel}{sPLS-DA model computed and returned inside
runSPLSDAmodels_Diablo() results.}

\item{dataTest}{List of new omic data sets and samples for same omics that
training to test model performances. Returned by splitDatatoTrainTest().}
}
\value{
Confusion matrix for sPLS-DA model
}
\description{
A diablo models assessing function.
}
\examples{
data("omic2", package = "multiSight")
splittedData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splittedData$data.train
data.test <- splittedData$data.test

diabloRes <- runSPLSDA(data.train)
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.
perfDiablo <- assessPerformance_Diablo(diabloModels, data.test)
perfDiablo$Omic1 # sPLS-DA's perf for omic1 data block.

}
