% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biosignerRes}
\alias{biosignerRes}
\title{Biosigner results}
\format{
An object of class \code{list} of length 2.
}
\source{
Returned by runSVMRFmodels_Biosigner() function.
}
\usage{
biosignerRes
}
\description{
Biosigner results list:
\describe{
  \item{model}{SVM and RF models}
  \item{biosignature}{Selected features of omic data sets}
}
}
\keyword{datasets}
