% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MLmodels_utils.R
\name{buildFeatTable}
\alias{buildFeatTable}
\title{Models util function}
\usage{
buildFeatTable(featVec, omicBlock, Y, deTable = NULL)
}
\arguments{
\item{featVec}{Selected features vector of one omic dataset.}

\item{omicBlock}{Omic dataset of selected features.}

\item{Y}{Omic sample classes.}

\item{deTable}{(optional) Deseq2 results table for an omic dataset.}
}
\value{
Returns table of features selected by classification model and
relative values.
}
\description{
To build detailed feature tables outputs for selected ones by 
machine learning models. Relative mean values by label class then 
log2FoldChange and p.adj values if DESeq2 have been computed are indicated.
}
\examples{
data("omic2", package = "multiSight")
splittedData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splittedData$data.train
data.test <- splittedData$data.test

#diabloRes <- runSPLSDA(data.train)
data("diabloRes", package = "multiSight")
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.
diabloFeatTable <- buildFeatTable(diabloFeats[[1]], 
  omic2[[1]], 
  omic2$Y)
diabloFeatTable

}
