% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_understand_utils.R
\name{convertToEntrezid}
\alias{convertToEntrezid}
\title{Understand utils function}
\usage{
convertToEntrezid(featList, fromDbList, organismDb)
}
\arguments{
\item{featList}{Feature lists from each omic data block.}

\item{fromDbList}{Database names vector. One by omic block (e.g. for 2 omics 
list(omic1 = "SYMBOL", omic2 = "ENSEMBL")). Returned by getDbFromInput() 
for app.}

\item{organismDb}{Organism database to convert features.}
}
\value{
featConverted
}
\description{
To convert features to entrezid to enrich.
}
\examples{
\donttest{
library(org.Mm.eg.db) # Organism's database
data("omic2", package = "multiSight")
splittedData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splittedData$data.train
data.test <- splittedData$data.test

diabloRes <- runSPLSDA(data.train)
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.
id_db <- list(omic1 = "ENSEMBL", omic2 = "ENSEMBL")
convFeat <- convertToEntrezid(diabloFeats, id_db, "org.Mm.eg.db")
}

featList <- list(Omic1 = c("ENSMUSG00000039621", 
                           "ENSMUSG00000038733", 
                           "ENSMUSG00000062031"), 
                 Omic2 = c("ENSMUSG00000031170", 
                           "ENSMUSG00000077495", 
                           "ENSMUSG00000042992"))
dbList <- list(Omic1 = "ENSEMBL",
                 Omic2 = "ENSEMBL")

convFeat <- convertToEntrezid(featList, dbList, "org.Mm.eg.db")

}
