% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deseqRes}
\alias{deseqRes}
\title{multiSight results}
\format{
An object of class \code{list} of length 2.
}
\source{
Returned by multiSight runMultiDeseqAnalysis() function
}
\usage{
deseqRes
}
\description{
DESeq2 results object obtained with several omic data sets.
}
\keyword{datasets}
