% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabloRes}
\alias{diabloRes}
\title{Diablo results}
\format{
An object of class \code{list} of length 3.
}
\source{
Returned by runSPLSDA() function.
}
\usage{
diabloRes
}
\description{
Diablo results list:
\describe{
  \item{design}{Covariance matrix design to maximize}
  \item{model}{sPLS-DA model}
  \item{biosignature}{Selected features of omic data sets}
}
}
\keyword{datasets}
