% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrichResList}
\alias{enrichResList}
\title{multiSight results}
\format{
An object of class \code{list} of length 2.
}
\source{
Returned by multiSight runMultiEnrichment() function
}
\usage{
enrichResList
}
\description{
enrichRes objects obtained with several omic data sets.
}
\keyword{datasets}
