% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_understand_fct_enrichment.R
\name{enrichWP}
\alias{enrichWP}
\title{Util function From clusterProfiler}
\usage{
enrichWP(gene, organism, pAdjustMethod, minGSSize, maxGSSize)
}
\arguments{
\item{gene}{Features to enrich.}

\item{organism}{Value chosen by user in Home tab..}

\item{pAdjustMethod, minGSSize, maxGSSize}{Numeric values chosen by user in
Biological Insights tab.}
}
\value{
enrichResult object with wikiPAthways database used.
}
\description{
Downloads wikiPathways data and computes ORA analysis with
provided features.
}
