% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hypothesisGenerator_fct_NetworkInference.R
\name{getDataSelectedFeatures}
\alias{getDataSelectedFeatures}
\title{Select only data values from diablo features selected.}
\usage{
getDataSelectedFeatures(matrixDataList, featureList = NULL)
}
\arguments{
\item{matrixDataList}{A matrix list according to omic type with
sample ID as columns and features as rows.}

\item{featureList}{A features list for each omic type.}
}
\value{
A concatenated matrix of all omic selected features.
}
\description{
Select only data values from diablo features selected.
}
\examples{
data("omic2", package = "multiSight")
splitData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splitData$data.train
data.test <- splitData$data.test

#diabloRes <- runSPLSDA(data.train)
data("diabloRes", package = "multiSight")
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.
omicMatrices <- getDataSelectedFeatures(omic2, diabloFeats)

}
