% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hypothesisGenerator.R
\name{mod_hypothesisGenerator_server}
\alias{mod_hypothesisGenerator_server}
\title{hypothesisGenerator Server Function}
\usage{
mod_hypothesisGenerator_server(
  input,
  output,
  session,
  obj,
  biosignature,
  featMethod
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{obj}{R6 object to wrap all data from different analysis.}

\item{biosignature}{Discriminant biological features selected.}

\item{featMethod}{character string to specify feature selection method 
(e.g. "diablo");}
}
\value{
Network inferences and PubMed queries results.
}
\description{
hypothesisGenerator Server Function
}
