% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_understand.R
\name{mod_understand_server}
\alias{mod_understand_server}
\title{understand Server Function}
\usage{
mod_understand_server(input, output, session, startSignal)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{startSignal}{input$start from Start button in Home tab.}
}
\value{
Biological insights from databases chosen by user by hypergeomtric 
tests (ORA) on DESeq2 or diablo (sPLS-DA) selected features.
}
\description{
understand Server Function
}
