% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_understand.R
\name{mod_understand_ui}
\alias{mod_understand_ui}
\title{understand UI Function}
\usage{
mod_understand_ui(id)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}
}
\value{
Displays UI output for Biological Insigths tab.
}
\description{
A shiny Module.
}
