% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_user_input.R
\name{mod_user_input_server}
\alias{mod_user_input_server}
\title{user_input Server Function}
\usage{
mod_user_input_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for shiny}
}
\value{
Launches multi-omic data analysis and saves results.
}
\description{
user_input Server Function
}
