% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hypothesisGenerator_fct_NetworkInference.R
\name{mutualInformationNI}
\alias{mutualInformationNI}
\title{Mutual Information network inference}
\usage{
mutualInformationNI(concatenatedMatrix, valueThreshold)
}
\arguments{
\item{concatenatedMatrix}{A concatenated matrix of all omic selected 
features. Returned by getDataSelectedFeatures().}

\item{valueThreshold}{Mutual Information Value threshold to select relevant 
values}
}
\value{
Each result for a mutual information network of selected features 
according to threshold values.
}
\description{
Mutual Information network inference
}
\examples{
data("omic2.rda")
splittedData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splittedData$data.train
data.test <- splittedData$data.test

diabloRes <- runSPLSDA(data.train)
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.
omicMatrices <- getDataSelectedFeatures(omic2, diabloFeats)
mutualInformationNI(omicMatrices, 0.8)

}
