% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omic2}
\alias{omic2}
\title{Multi-omic data with 2 omics.}
\format{
A list of 2 dataframes and 1 factor vector 30 rows:
\describe{
  \item{rnaRead}{transcriptomic simulated data}
  \item{dnaRead}{genomic simulated data}
  \item{Y}{30 samples' classes}
}
}
\source{
MOSim package used to simulate omic data.
}
\usage{
omic2
}
\description{
A dataset containing simulated multi-omic data with 30 samples.
}
\keyword{datasets}
