% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hypothesisGenerator_fct_NetworkInference.R
\name{partialCorrelationNI}
\alias{partialCorrelationNI}
\title{Partial Correlation network inference}
\usage{
partialCorrelationNI(concatenatedMatrix, valueThreshold)
}
\arguments{
\item{concatenatedMatrix}{A concatenated matrix of all omic selected 
features. Returned by getDataSelectedFeatures().}

\item{valueThreshold}{Partial Correlation Value threshold to select 
relevant values}
}
\value{
Each result for a partial correlation network of selected features 
according to threshold values.
}
\description{
Partial Correlation network inference
}
\examples{
data("omic2", package = "multiSight")
splitData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splitData$data.train
data.test <- splitData$data.test

#diabloRes <- runSPLSDA(data.train)
data("diabloRes", package = "multiSight")
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.
omicMatrices <- getDataSelectedFeatures(omic2, diabloFeats)
partialCorrelationNI(omicMatrices, 0.8)

}
