% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hypothesisGenerator_fct_PubmedInsight.R
\name{pubmedInsight}
\alias{pubmedInsight}
\title{Select only data values from diablo features selected.}
\usage{
pubmedInsight(session, input, output, query, featMethod)
}
\arguments{
\item{input, output, session}{Internal parameters for shiny}

\item{query}{character vector to query pubmed database}

\item{featMethod}{character string to specify 
feature selection method (e.g. "diablo")}
}
\value{
uiOutput, a html content to embed in client side.
}
\description{
Select only data values from diablo features selected.
}
