% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MLmodels_utils.R
\name{runFeatureNumberTuning}
\alias{runFeatureNumberTuning}
\title{diablo util function}
\usage{
runFeatureNumberTuning(dataTrainList, YClassVector, ncomp, design)
}
\arguments{
\item{dataTrainList}{List of data block training part.}

\item{YClassVector}{List of your sample classes vector provided}

\item{ncomp}{Component number in splsda model obtained from 
runComponentNumberTest function launched by runSPLSDAmodels_Diablo 
function.}

\item{design}{Covariance matrix design obtained from}
}
\value{
Grid of number of features to select by sPLS-DA model for each 
component and omic.
}
\description{
A feature number tuning function of splsda model
for each component of each data block.
}
