% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_understand_fct_enrichment.R
\name{runMultiDeseqAnalysis}
\alias{runMultiDeseqAnalysis}
\title{Enrichment function}
\usage{
runMultiDeseqAnalysis(omicDataList, padjUser)
}
\arguments{
\item{omicDataList}{List of omic data blocks with Y class vector.}

\item{padjUser}{Threshold for p-value adjusted to select features
according to padj values in DESeq2 table.}
}
\value{
List of DESeq2's Differential Expression tables for all omic 
datasets (e.g. BaseMean, Log2FoldChange, padj columns).
}
\description{
Runs DESEQ2 analysis on all omic data blocks.
}
\examples{
data("omic2", package = "multiSight")
deseqRes <- runMultiDeseqAnalysis(omic2, 0.05)

}
