% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_understand_fct_enrichment.R
\name{runMultiEnrichment}
\alias{runMultiEnrichment}
\title{Main understand module function}
\usage{
runMultiEnrichment(
  omicSignature,
  databasesChosen,
  organismDb,
  pvAdjust = "BH",
  minGSSize = 5,
  maxGSSize = 800,
  pvStouffer = 0.1
)
}
\arguments{
\item{omicSignature}{Feature lists from each omic data block.}

\item{databasesChosen}{Which biological database 
c(reactome, kegg, wikiPathways, MF, CC, BP)}

\item{organismDb}{Organism provided by user in Home tab.}

\item{pvAdjust}{pv adjust method (e.g "BH" for Benjamini-Hochberg)}

\item{minGSSize, maxGSSize, pvStouffer}{Numeric values chosen by user 
in ui.}
}
\value{
Wraps in obj enrichment results for all databases and all omics.
}
\description{
Launches functional enrichment of features provided
for every databases furnished. Run by utils function runMultiOmicEnrichment
}
\examples{
\donttest{
data("omic2", package = "multiSight")
splittedData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splittedData$data.train
data.test <- splittedData$data.test

diabloRes <- runSPLSDA(data.train)
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.
id_db <- list(omic1 = "ENSEMBL", omic2 = "ENSEMBL")
}
library(org.Mm.eg.db, warn.conflicts = FALSE) #Organism's database
featList <- list(Omic1 = c("ENSMUSG00000039621", 
                           "ENSMUSG00000038733", 
                           "ENSMUSG00000062031"), 
                 Omic2 = c("ENSMUSG00000031170", 
                           "ENSMUSG00000077495", 
                           "ENSMUSG00000042992"))
dbList <- list(Omic1 = "ENSEMBL",
                 Omic2 = "ENSEMBL")
convFeat <- convertToEntrezid(featList, dbList, "org.Mm.eg.db")

## To enrich features
database <- c("reactome", "MF")
enrichTables <- runMultiEnrichment(databasesChosen = database,
                                   omicSignature = convFeat,
                                   organismDb = "org.Mm.eg.db")
  
}
