% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MLmodels_fct_diablo.R
\name{runSPLSDA}
\alias{runSPLSDA}
\title{MLmodels diablo function}
\usage{
runSPLSDA(dataTrain)
}
\arguments{
\item{dataTrain}{Data train set to build classification models.
Returned by splitDatatoTrainTest().}
}
\value{
Returns sPLS-DA model, performances, features selected.
}
\description{
A model and selection features function from mixOmics.
}
\examples{
splittedData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splittedData$data.train
data.test <- splittedData$data.test

diabloRes <- runSPLSDA(data.train)
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.

}
