% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MLmodels_fct_biosigner.R
\name{runSVMRFmodels_Biosigner}
\alias{runSVMRFmodels_Biosigner}
\title{MLmodels biosigner function}
\usage{
runSVMRFmodels_Biosigner(dataTrain)
}
\arguments{
\item{dataTrain}{Data train set to build classification models. 
Returned by splitDatatoTrainTest(). List of Omic blocks and Y class vector.}
}
\value{
Models and features selected for each omic block in dataTrain.
}
\description{
A model and selection features function from biosigner.
}
\examples{
splittedData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splittedData$data.train
data.test <- splittedData$data.test

biosignerRes <- runSVMRFmodels_Biosigner(data.train)
biosignerModels <- biosignerRes$model #list of SVM/RF models for each omic.
biosignerFeats <- biosignerRes$biosignature #selected features for each omic.

}
