% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MLmodels_utils.R
\name{splitDatatoTrainTest}
\alias{splitDatatoTrainTest}
\title{MLmodels util function}
\usage{
splitDatatoTrainTest(MultiOmicData, freq = 0.8)
}
\arguments{
\item{MultiOmicData}{list of your data blocks}

\item{freq}{Split proportion of train samples}
}
\value{
Return two data sets: first to train model and second to assess it.
}
\description{
To split list of omic data in data train and data test subsets.
}
\examples{
data("omic2", package = "multiSight")
splitData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splitData$data.train
data.test <- splitData$data.test

}
