% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_understand_fct_enrichment.R
\name{stoufferTable}
\alias{stoufferTable}
\title{Stouffer's p-value computing with all pvalues from all enrichment
tables provided and all gene sets to build enrichment maps.}
\usage{
stoufferTable(enrichmentResult)
}
\arguments{
\item{enrichmentResult}{enrichResults@result list.}
}
\value{
enrichment table results merged with Stouffer's p-value
non-weighted and weighted.
}
\description{
Stouffer's p-value computing with all pvalues from all enrichment
tables provided and all gene sets to build enrichment maps.
}
\examples{
data(enrichResList, package = "multiSight")
enrichResList # list of enrichRes objects (e.g. enrichKEGG() results)
multiOmicRes <- stoufferTable(enrichResList)
multiOmicRes$table # table with stouffer's values  
multiOmicRes$moEnrichRes # enrichRes object for clusterProfiler plots  

\donttest{
data("omic2", package = "multiSight")
splittedData <- splitDatatoTrainTest(omic2, 0.8)
data.train <- splittedData$data.train
data.test <- splittedData$data.test

diabloRes <- runSPLSDA(data.train)
diabloModels <- diabloRes$model #sPLS-DA model using all omics.
diabloFeats <- diabloRes$biosignature #selected features for each omic.
id_db <- list(omic1 = "ENSEMBL", omic2 = "ENSEMBL")
library(org.Mm.eg.db) # Organism's database
convFeat <- convertToEntrezid(diabloFeats, id_db, "org.Mm.eg.db")
database <- c("reactome", "MF")
#enrichTables <- runMultiEnrichment(databasesChosen = database,
#  omicSignature = convFeat,
#  organismDb = "org.Mm.eg.db")
# enrichmentTables <- enrichTables$pathways$reactome$enrichObj
#enrichResList # list of enrichRes objects (e.g. enrichKEGG() results)
data(enrichResList, package = "multiSight")
multiOmicTable <- stoufferTable(enrichResList)
}

}
